pyproject.toml
[project]
name = "fastsecure"
version = "0.1.0"
description = "Flexible authentication system for FastAPI applications"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Igor Benav", email = "igor.magalhaes.r@gmail.com" }
]
keywords = ["fastapi", "authentication", "oauth", "jwt", "session"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.4",
]
